Final Commit.
This is the final commit as of 31st August 2024, with the final version being 1.0.0. No further developments are planned for the project. If in case any future changes are required, it will be committed via pull request. Current version can be either recreated, while a public version is readily available [here](https://ingen.only-fans.club).